Optimization Hint Flag in the DM schema - hasLargeList
When storing a collection-type value with a large list, it is not always
good to treat it as a nested table if the nested table is narrow and
dense.  A better option could be store it in the row-based format, which
is hinted by a new attribute 'hasLargeList' of the field.

The implementation would choose whether and how to deal with this hint.
In other words, it is optional in the implementation.