Fix a bug with WritableRecord
In WritableRecord, there is an object array that contains the value or
update of each attribute in the record.

The length of the array is defined when the record is stored, which may
not be aligned with the CURRENT metadata. This can cause an exception if
the OLD record is updated.

The fix will make the length of the array based on the CURRENT table
schema.